models/persona_training/train.py
# Persona-Based AI Training
# ==========================
#
# This module trains AI agents based on persona filling logic.
# Each persona becomes a specialized agent for filling their document types.

"""
Persona Training Pipeline

Trains AI models to replicate how specific personas (user types) fill documents.

Usage:
    python -m models.persona_training.train --persona hr-professional --documents w4,i9
"""

import argparse
import json
import logging
from pathlib import Path
from typing import Dict, Any, List, Optional
from dataclasses import dataclass, asdict

import yaml

logger = logging.getLogger(__name__)


@dataclass
class PersonaConfig:
    """Configuration for a persona."""
    id: str
    name: str
    document_expertise: Dict[str, List[str]]
    data_sources: List[Dict[str, Any]]
    knowledge: Dict[str, List[str]]
    training_config: Dict[str, Any]


@dataclass
class FillingLogic:
    """Filling logic for a specific document."""
    document: str
    field_mappings: List[Dict[str, Any]]
    decision_logic: List[Dict[str, Any]]
    validations: List[Dict[str, Any]]


@dataclass
class TrainingExample:
    """Single training example."""
    id: str
    scenario: str
    input_data: Dict[str, Any]
    filled_form: Dict[str, Any]
    notes: Optional[str] = None


def load_persona(persona_dir: Path) -> PersonaConfig:
    """Load persona configuration from directory."""
    persona_file = persona_dir / "persona.yaml"
    
    with open(persona_file, 'r') as f:
        data = yaml.safe_load(f)
    
    return PersonaConfig(
        id=data['id'],
        name=data['name'],
        document_expertise=data.get('document_expertise', {}),
        data_sources=data.get('data_sources', []),
        knowledge=data.get('knowledge', {}),
        training_config=data.get('training_config', {})
    )


def load_filling_logic(persona_dir: Path, document: str) -> Optional[FillingLogic]:
    """Load filling logic for a specific document."""
    logic_file = persona_dir / "filling-logic" / f"{document}-logic.yaml"
    
    if not logic_file.exists():
        logger.warning(f"No filling logic found for {document}")
        return None
    
    with open(logic_file, 'r') as f:
        data = yaml.safe_load(f)
    
    return FillingLogic(
        document=data['document'],
        field_mappings=data.get('field_mappings', []),
        decision_logic=data.get('decision_logic', []),
        validations=data.get('validations', [])
    )


def load_training_examples(persona_dir: Path, document: str) -> List[TrainingExample]:
    """Load training examples for a document."""
    examples_file = persona_dir / "training" / "examples" / f"{document}-examples.jsonl"
    
    if not examples_file.exists():
        logger.warning(f"No training examples found for {document}")
        return []
    
    examples = []
    with open(examples_file, 'r') as f:
        for line in f:
            data = json.loads(line.strip())
            examples.append(TrainingExample(
                id=data['id'],
                scenario=data.get('scenario', ''),
                input_data=data.get('input', {}),
                filled_form=data.get('filled_form', {}),
                notes=data.get('notes')
            ))
    
    return examples


def create_training_dataset(
    persona: PersonaConfig,
    filling_logic: FillingLogic,
    examples: List[TrainingExample]
) -> Dict[str, Any]:
    """
    Create training dataset for the model.
    
    Combines:
    - Persona context (role, knowledge, data sources)
    - Filling logic (mappings, decisions, validations)
    - Training examples (input → output pairs)
    """
    
    # Build context for the model
    context = {
        "persona": {
            "id": persona.id,
            "name": persona.name,
            "knowledge": persona.knowledge
        },
        "document": filling_logic.document,
        "field_mappings": filling_logic.field_mappings,
        "decision_logic": filling_logic.decision_logic,
        "validations": filling_logic.validations
    }
    
    # Format examples for training
    training_data = []
    for example in examples:
        training_data.append({
            "input": {
                "context": context,
                "data": example.input_data
            },
            "output": example.filled_form,
            "scenario": example.scenario
        })
    
    return {
        "persona_id": persona.id,
        "document": filling_logic.document,
        "num_examples": len(training_data),
        "examples": training_data
    }


def train_persona_model(
    dataset: Dict[str, Any],
    output_dir: Path,
    config: Dict[str, Any]
) -> None:
    """
    Train the persona model.
    
    This can use various approaches:
    1. Fine-tune an LLM on the examples
    2. Train a specialized extraction + filling model
    3. Create a rule-based system from the logic
    """
    
    persona_id = dataset['persona_id']
    document = dataset['document']
    
    logger.info(f"Training {persona_id} persona for {document}")
    logger.info(f"Training examples: {dataset['num_examples']}")
    
    # Create output directory
    model_dir = output_dir / persona_id / document
    model_dir.mkdir(parents=True, exist_ok=True)
    
    # Save training dataset
    with open(model_dir / "training_data.json", 'w') as f:
        json.dump(dataset, f, indent=2)
    
    # TODO: Implement actual model training
    # Options:
    # 1. Fine-tune LLM (GPT-4, Claude, Llama, etc.)
    # 2. Train encoder-decoder model
    # 3. Compile rules into executable logic
    
    logger.info("Model training placeholder - implement based on chosen approach")
    
    # Save model config
    model_config = {
        "persona_id": persona_id,
        "document": document,
        "training_examples": dataset['num_examples'],
        "accuracy_target": config.get('accuracy_target', 0.95)
    }
    
    with open(model_dir / "model_config.json", 'w') as f:
        json.dump(model_config, f, indent=2)
    
    logger.info(f"Model artifacts saved to {model_dir}")


def main():
    parser = argparse.ArgumentParser(description='Train persona-based filling model')
    parser.add_argument('--persona', type=str, required=True,
                        help='Persona ID to train (e.g., hr-professional)')
    parser.add_argument('--documents', type=str, required=True,
                        help='Comma-separated document types (e.g., w4,i9)')
    parser.add_argument('--output', type=str, default='models/trained',
                        help='Output directory for trained models')
    parser.add_argument('--personas-dir', type=str, default='personas',
                        help='Directory containing persona definitions')
    args = parser.parse_args()
    
    logging.basicConfig(level=logging.INFO)
    
    # Load persona
    persona_dir = Path(args.personas_dir) / args.persona
    if not persona_dir.exists():
        logger.error(f"Persona not found: {args.persona}")
        return
    
    persona = load_persona(persona_dir)
    logger.info(f"Loaded persona: {persona.name}")
    
    # Train for each document
    documents = [d.strip() for d in args.documents.split(',')]
    
    for document in documents:
        logger.info(f"\n--- Training for {document} ---")
        
        # Load filling logic
        logic = load_filling_logic(persona_dir, document)
        if not logic:
            continue
        
        # Load examples
        examples = load_training_examples(persona_dir, document)
        if not examples:
            logger.warning(f"No examples for {document}, skipping")
            continue
        
        logger.info(f"Loaded {len(examples)} training examples")
        
        # Create dataset
        dataset = create_training_dataset(persona, logic, examples)
        
        # Train model
        train_persona_model(
            dataset,
            Path(args.output),
            persona.training_config
        )
    
    logger.info("\nTraining complete!")


if __name__ == '__main__':
    main()
