models/requirements.txt
# AI Model Training Dependencies
# ================================

# Core ML frameworks
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0

# Document processing
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.0.0
pymupdf>=1.23.0

# Layout analysis
layoutparser>=0.3.4
detectron2>=0.6

# NLP
spacy>=3.7.0
sentence-transformers>=2.2.0

# Training utilities
accelerate>=0.24.0
wandb>=0.16.0
tensorboard>=2.15.0

# Evaluation
scikit-learn>=1.3.0
seqeval>=1.2.2
editdistance>=0.6.2

# Data handling
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
jsonlines>=4.0.0

# Export
onnx>=1.15.0
onnxruntime>=1.16.0

# Development
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
